Refactor `useOnlineHeatmapDataService`
Extracted the logic from the extremely long `useOnlineHeatmapDataService` function into two new separate, specific hooks:
1. `useHeatmapTask`: Manages the HeatmapTask querying, polling, and related states (`zVisible`, etc.).
2. `useHeatmapDataFetchers`: Contains all the general data fetching utilities (maps, logs, sessions, players, etc.) mapped directly to API routes.

This improves maintainability, readability, and makes the heatmap services easier to mock and test independently.
